refactor: songs view
